requirements.txt
fastapi==0.99.1
uvicorn==0.23.1
python-whois==0.7.3
validators==0.20.0
httpx==0.24.1
Jinja2==3.1.2
python-multipart==0.0.6